move layout components into own directory
