+++ src/components/layout/DashHeader.vue
@@ -46,10 +46,3 @@ export default {
   }
 }
 </script>
-
-<style lang="scss" scoped>
-.navbar-nav > .messages-menu > .dropdown-menu > li .menu > li > a > h4,
-.navbar-nav > .messages-menu > .dropdown-menu > li .menu > li > a > p {
-  margin-left: 0;
-}
-</style>
