+++ src/components/layout/SidebarMenu.vue
